+++ lib/ios/ThaliCore/ThaliCoreTests/AppStateNotificationsManagerTests.swift
@@ -3,7 +3,8 @@
 //  AppStateNotificationsManagerTests.swift
 //
 //  Copyright (C) Microsoft. All rights reserved.
-//  Licensed under the MIT license. See LICENSE.txt file in the project root for full license information.
+//  Licensed under the MIT license. See LICENSE.txt file in the project root for full license
+//  information.
 //
