+++ lib/ios/ThaliCore/ThaliCoreTests/BrowserManagerTests.swift
@@ -3,7 +3,8 @@
 //  BrowserManagerTests.swift
 //
 //  Copyright (C) Microsoft. All rights reserved.
-//  Licensed under the MIT license. See LICENSE.txt file in the project root for full license information.
+//  Licensed under the MIT license. See LICENSE.txt file in the project root for full license
+//  information.
 //
 
 import XCTest
